# Conflicts:
#	src/sample/Controller.java

+++ src/sample/Point.java
@@ -1,15 +1,15 @@
 package sample;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y){
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX(){
+    public double getX(){
         return x;
     }
 
@@ -17,7 +17,7 @@ public class Point {
         this.x = x;
     }
 
-    public int getY(){
+    public double getY(){
         return y;
     }
